chore: remove fs usage from dendron-viz
