+++ packages/dendron-cli/src/commands/visualizeCLICommand.ts
@@ -1,7 +1,10 @@
-import { generateSVG, InputArgs } from "@dendronhq/dendron-viz";
+import { InputArgs, getVisualizationMarkup } from "@dendronhq/dendron-viz";
 import { Argv } from "yargs";
 import { CLICommand, CommandCommonProps } from "./base";
 import { setupEngine, setupEngineArgs, SetupEngineResp } from "./utils";
+import * as fs from "fs-extra";
+import { VaultUtils } from "@dendronhq/common-all";
+import path from "path";
 
 type CommandOpts = InputArgs & SetupEngineResp & CommandCommonProps;
 
@@ -28,7 +31,29 @@ export class VisualizeCLICommand extends CLICommand {
   }
 
   async execute(opts: CommandOpts) {
-    await generateSVG(opts);
+    const { out, engine, wsRoot } = opts;
+
+    /* Ensure that the provided directory exists. If not present, this creates the directory */
+    if (out) await fs.ensureDir(out);
+
+    if (!engine) throw new Error("Engine is not initialized");
+
+    await Promise.all(
+      engine.vaults.map(async (vault) => {
+        /* Get vault name */
+        const vaultName = VaultUtils.getName(vault);
+        const outputFile = path.join(out || wsRoot, `diagram-${vaultName}.svg`);
+        /* Create React component for visualization */
+        const html = await getVisualizationMarkup({
+          ...opts,
+          vault,
+          notes: engine.notes,
+        });
+        /* Write markup to the output file */
+        await fs.writeFile(outputFile, html);
+      })
+    );
+
     return { exit: true };
   }
 }
